Update login page to add icons to auth providers
Icons fetched from: https://github.com/intridea/authbuttons

I believe that this use of the copyrighted / trademarked logos probably
qualifies as fair use. The logos are affixed to buttons that link directly to
the owners' respective sites, and are used to perform the authentication
service that those sites provided for public use.