Add support for login via Brown University Shibboleth
This depends upon mod_shib in Apache, including a site-specific
Shibboleth Service Provider configuration file, which cannot all be
committed for security reasons. For this reason, the Shibboleth login
functionality will require modification to be used at other installations.
Regardless, a future task is to remove all Brown University-specific
information from the Shibboleth login.

The Brown University logo is taken from:
http://en.wikipedia.org/wiki/File:BrownU-CoA.svg
but fetched as a 32-pixel PNG.

The logo likely qualifies as fair use, for the same reasons as the
Facebook and Google logo, and the fact that Brown University has been
cooperating with us to set up the system.