Removed the option to specify an authentication database
MongoEngine does not currently support the ability to authenticate against
a different database than the one being used. A pull request to add this
feature has been submitted upstream, but not yet merged or released.

Until MongoEngine properly supports this functionality, it should be
replicated in the local app.

In the very short term, until the functionality is added locally, it will
be removed entirely to prevent PyMongo from throwing exceptions from
unknown arguments.