Update Apache WSGI to suggest setting an path to the site config file
After extensive research, I can't find any way to change the environment
of the WSGI daemon before it imports the WSGI site file for the first
time. Apache's "SetEnv" directive only applies to the requests that are
provided to the site.

An alternative location for placing site-specific configuration files
outside of the code would be Flask's instance directory. While potentially
more appropriate for an installed app, it is less flexible than the
environment variable approach.