Fix a bug in MultipleDatabaseModelDocument
An instance of a MultipleDatabaseModelDocument-derived type could forget
its Mongo collection name under some circumstances. The MongoEngine
metaclasses assume that a newly created object's class metainformation is
stored in 'meta'. The metaclasses then move the metainformation to '_meta'
for the object instance. This moves the metainformation from '_meta' back
to 'meta' when copying a MultipleDatabaseModelDocument.