Create all model mutable default values using a callable
MongoEngine default values are similar to Python function default
arguments, in that only a single instance of the default value is created
when the MongoEngine field is first defined. Subsequent requests for the
default value would return the same mutable object, unless it's a callable
that creates a new mutable object on demand.