Add unique IDs to Slide-Atlas menu buttons
Add the capability to specify the ID of each type of  button which is
found in the Slide-Atlas UI.

Add a test to test for the correct number of available buttons and ensures
that each button generates a visible menu.

The test uses Selenium and Python.  it only requires one argument which is
passed with the '-r' flag.  That argument is the webroot of the
Slide-Atlas instance with the trailing slash.
For example
  python test_glViewer_menus.py -r "http://localhost:8080/"

The test will open the "HD001 Renal Infarct - Coagulative Necrosis" view