Add to Slide-Atlas Selenium tests
Change the page to be loaded for each test to be one of the images
available for all users, signed in or not.

Add file with interaction for the navigation canvas testing the rotation
of the widget and add the skeleton of a test for changing the size of the
canvas.

Add interactions with each menu button's content to attempt to exercise
each option within the popup menu.  Comment out the editTab content as it
seems to hide other elements.